Including guidance for CRL number handling seems out of scope
While Section 5 of RFC6487 states that "[w]here two or more CRLs are
issued by the same CA, the CRL with the highest value of the 'CRL
Number' field supersedes all other CRLs issued by this CA"; yes, but...

Section 8 of RFC 6487 clarifies:

  CRL Scope and Key Values:
     For CRL Scope, this profile specifies that a CA issues a *single*
     CRL at a time, and the scope of the CRL is all certificates
     issued by this CA.  Because the CA instance is bound to a
     single key pair, this implies that the CA's public key, the key
     used to validate the CA's CRL, and the key used to validate the
     certificates revoked by that CRL are all the same key value.

My reading is that the CRLDP indeed can contain multiple locations,
of the *same* instance of the CRL (for example, one might want to
distribute the same CRL via rsync and HTTPS).

Therefor, as only a single CRL file is to be expected in each
Manifest FileNameAndHash listing. That one CRL will be the applicable
CRL, regardless of its CRL number. So to unbreak a situation where
the CRLNumber maxed out, just use a new filename. (But, to me,
discussing this mechanism in great detail seems out of scope for this
internet-draft)